New deal! The server handles everything, it parses in a (for testing) config.xml file which contains default configuration properties for the game, it handles the creation of the game and firing up of the standalone Flash player for whatever number of clients are desired:
ant test             - starts the server and one Flash client
ant -Dplayers=2 test - starts the server and two Flash clients

This will allow us to, in the future, create a test server that can be used by
FAT developers wherein it will pop up a little UI that they can use to adjust
their game configuration and set the number of players and tell it where their
target SWF is and then click a "test" button and it will do all the necessary
magic and they'll never have to touch the command line.

We could even create a UI that allows them to add new configuration parameters
so that they don't even need to create a config.xml and when they're done they
could click a button and it could give them the XML they need to use when
uploading their game to Whirled. Hotness.