Our first player-contributed source!  I haven't tested these, but have read over them and ensured that they compile.  Also, the works of the author (jdnx429) use these classes and are well-tested on Whirled.  Submitted with minor changes: mostly code formatting.
From jdnx429:

"Alrighty, I have a couple of AS files that have lots of interesting functions that I've been using.
The functions in the ColorTools have been quite helpful in creating some of the controllers for
EveryColor and of course, Paint Factory.  They PhysicalBase is the base class that I used for the
fireworks I have listed in the Catalog. I have another game waaay in the back burner that uses
subclasses of the PhysicalBase for moving things around. It's helpful for quick and dirty physics."