Rejigged class names to make more sense and preserve the linear inheritance chain for whirled -> (toybox) -> msoy classes
The new hierarchy is now consistent for *Controller, *Config and *Backend:

  GameConfig -> BaseGameConfig -> WhirledGameConfig -> toybox -> msoy
                               -> ThaneGameConfig

  GameController -> BaseGameController -> WhirledGameController -> ...
                                       -> ThaneGameController

  BaseGameBackend -> WhirledGameBackend
                  -> ThaneGameBackend

Essentially, we are committing to flash/java for the "whirled"ly named things
and merely giving thane a helping hand by sharing base classes. The previous
refactoring attempted to somewhat abstract the notion of a "whirled" controller
or config, but this turned out to be very dangerous territory since many places
assume that whirled exposed a concrete, more or less fixed implementation
(toybox, game gardens and msoy).

For reference, the first pass refactoring created this hierarchy:

  ... -> WhirledGameConfig -> FlashWhirledGameConfig -> MsoyGameConfig
                           -> ThaneWhirledGameConfig
                           -> ToyBoxGameConfig

  ... -> WhirledGameController -> FlashWhirledGameController -> msoy
                               -> ThaneWhirledGameController

  BaseGameBackend -> FlashGameBackend
                  -> ThaneGameBackend