Unrolled ThaneGameController inheritance * whirled BaseGameController functions and fields * vilya GameController functions and fields * bilya TurnGameController functions and fields * not PlaceController
Explanation:

A game controller is designed to be a place controller. Thane clients are
simply not able to use this, mostly because they host multiple agents, each
assigned to its own game. In order to make thane clients work as place
controllers, mutliple clients would have to run on the same connection or at
least in the same process, a complex scenario to say the least.

The possibility of faking the place-ness of a thane controller was
investigated. But in order to use a game controller, it must be init'ed with a
crowd context. Thane does not have one of those. If we provide a fake one,
the place controller (and the game controller) will crash when the fields are
accessed. If we try to provide a real one, we'd need to subclass all the
services and shunt the inappropriate behavior. This would be far messier.

Lastly, TurnGameController had to be unrolled because it is only usable with
GameController (there is an unchecked cast in the constructor).