Cleanup the magic "server agent" numbers plus other msoy-commits feedback
Regarding the magic numbers, rather than waffle around on the mailing list,
I'll give my observations here that led me to this approach:

* Most coders are perfectly at ease with "0 means X". This can even be
stretched to "-1 means Y". Any more than that becomes annoying and better
abstractions are needed. Case in point, having to use int.MIN_VALUE was what
kicked off the magic number discussions. This is clearly a non-memorable value.

* It will be very rare to actually need to type "ALL_PLAYERS" when calling
sendMessage, deal or pick. You'd simply leave the parameter off. That's why
it's last.

* Default parameter values are part of the documenation, even if no one will
normally need to type them. Having "ALL_PLAYERS" appear there enhances the
readability of the docs.

* Shortcut methods are much more robust and usable than constants. Shortcuts
like 'sendMessageToAgent' or 'isFromServerAgent' mean that users are not
forced to import a *SubControl or GameConstants and introduce more points of
failure for their code.

Another thing worth mentioning is a shift in terminology. Referring to the
game's server side code as "server" was just too ambiguous. "Server agent"
is the new buzz word (unless it gets vetoed). The SDK can use this from
now on. It will be introduced in the wiki somewhere.

Some Rationale: The ease of use of the SDK is higher priority than keeping
all of our 0's, -1's and MIN_VALUES's in one place (it was bound to be an
imperfect anyway due to having it in 2 languages). Also, a constants class
is confusing because it binds together otherwise disparate information. You
wouldn't put all your literal strings in one file just because they were
strings, would you?