So, Java's built-in serialization is unbelievably verbose. We're switching to Streamables instead, and unless we've misunderstood the feature completely it looks like we can use the addTranslation() mechanism to save a lot of further overhead.
This is still kind of clunky, though, because the values we're encoding can only be byte[], byte[][], or GameMap instances. It seems possibly excessive to use a full serialization mechanism for values that are so close to already serialized. We *could* do something home-grown, with a marker byte, or whatnot.

However, in the interest of making the format concrete and not spending hours fiddling over a few bytes overhead, this is probably a decent compromise. Disagreements should be voiced sooner rather than later.