Instead of saying "Unable to call host code" and throwing an error, tell the creator that it's THEIR FUCKING FAULT, log everythihg, and just return undefined (which will throw an error anyway in many cases).
Mark my words: this will not be enough. I'll see a checkin in may
in the msoy codebase where someone "protects" against this.

I think throwing an error is the quickest way to tell
someone that they fucked up. If someone lists an avatar anyway
when it throws errors, that's _not_ our problem. You wouldn't
list an avatar with its head on backwards. Same thing. It's
broken, the creator needs to fix it, end of story.