Use the new sytem property set by project-include.xml to determine where to store the game's user cookies.  The next SDK release should include the following:
User cookies are now stored at a different location for each game that the developer is working
on.  Each game's build.xml should include the name parameter on the <project> tag.  The cookies
are the stored at userPrefs/gameCookieManager/<gameName> for each game.  They were previously stored
at userPrefs/gameCookieManager, so in order to keep the cookies for a currently in-development game,
they must be moved from the old location to the new location.

On Unix-based systems, Java stores user preferences in ~/.java/.userPrefs.  In Windows, they are
stored in the Registry, and should be found at JavaSoft/Prefs.