Fixed remixing to work in CS4. It took a little while to isolate, but the fix was pretty straightforward. It's just so fucking perfect that adobe seems to have two different AS3 compilers that do things differently. You'd think they'd be able to make a working compiler for a language they spec'd themselves.
(Part of the reason this took a while to track down was that toString()ing
an XMLList seemed to turn it into nothing, and I'm pretty sure that used
to just work. I was tracing things out, figuring out what was fucking up
and thought the error was happening earlier than it was.
Perhaps this is another compiler difference between CS4 and the rest.)