First pass at location-aware sound effects.  As advertised, they pan left/right and fade in volume depending on the location of the object causing the sound relative to the center of the screen.
DynamicSprite.SOUND_NORMAL_DISTANCE is the number of pixels over which a sound's volume will scale
and pan.  The square of area that includes objects capable of being heard is
SOUND_NORMAL_DISTANCE * 2 on a side.  Right now this is defined as Metrics.DISPLAY_HEIGHT * 2.  That
can probably be a lot smaller.  We'll see what the bosses think.

Another possible improvement would be to not scale sound volume and/or do panning linearly.  I could
see us possibly wanting to mess around with other functions for that scale, depending on how things
sound.  Probably this simple implementation will be enough.