I can't believe how long it took to (re)discover this issue.  This should have been changed when msoy's changed.
This does mean that game projects will now be targeting flash 10.  This shouldn't cause any issues,
as far as I know.  It also means that the kiddies are going to have to download a modern flex SDK.
That's probably a good thing - they should be keeping up to date anyway.