Another quick tweak based on Rayback.  I shouldn't just sum the times because later in VM life the difference between one object's times and another's will mean less than it does early in VM life. The reason I went with that approach to begin with though is still valid:  I want 3 accesses 5 seconds ago to mean more than 3 accesses 58 seconds ago.  So to fix: sum the difference of each time with the threshold so that VM life doesn't enter the equation.
Hrm...  I wonder if this means that this isn't strictly an LFU.  More of an LFU/LRU hybrid?  This
is an algorithm that I just plucked out of my head - maybe I'll go do more research on cache
expiration techniques when I have some spare time.  At the very least, I feel like I might be making
an assumption about what's best for this cache based on what I'm currently using it for that may
not hold true for all applications of a strict LFU.