Fix the cookie problem.
I broke it. How could I be so stupid! Jesus Fuck, can't I code my way out
of a wet paper back? (Trying to be as harsh on myself as I am on others.)

When doing the id switchover,  I changed setting a cookie to verify with
validateWritePermission(). Now- that method is wack.
I don't kow how *anything* works with that. It allows the agent to turn
a playerId into a body, but if called by a non-agent it ONLY works if
the passed-in playerId is 0. Wha?? This is duplicated in AVRG code.

We already call validateUser, which requires that the caller be a player or agent.
(Although, problematically, a watcher can be assigned as 'controller' of the game,
which should probably be changed.)
So, added a new method, validatePlayerId() that does what we want for cookies.
- Transform 0 into a proper id.
- Reject if the agent passes 0.
- For party games, ensure the id is for a player in the room.
- For seated games, ensure the id is for a player, EVEN IF ABSENT.