New APIs for spending bars directly. I decided to continue to route this through the item pack system, we will simply allow "approved" games to purchase and consume item packs without confirmation from the player.
This allows us to ensure that the bar transaction results in some persistent
token being delivered to the game. We'll either make both happen or neithre.
The game can then use that token to note that the player has some time, or turn
around and immediately consume it using consumeItemPack() if the purchase
represented something ephemeral.