context.allowLoadBytesCodeExecution = true;
Without this, a Flex app that uses SWFResource on a swf with executable code will throw a SecurityError: Error #3015.  It's rather annoying when errors don't automatically generate trace statements, leaving you with Google and guesswork and a million trace() statements to figure it out.

I still don't know what the executable code is.  Is it stuff left on the stage on an loaded swf?  Jon needs this done tonight, so committing now.